add install all CLI and Installer support
